Fix ifdefs to allow installation on Mac. Refs #11.
The src/Paths module, which is standard for Mac apps on Hails,
does not handle OSX, so it refers to functions that simply
do not exist.

This change modifies the src/Paths.hs module so that every ifdef for linux
includes darwin as well (which needs to be treated the same way).