Rename titan-gui to titan. Refs #16.
We already have a haskell package with the name titan, but none of the packages
in the repo match that name.

This change renames the package titan-gui to just titan. The files inside the
source tree are adapted due to the module Paths_titan, which is imported and
re-exported, being autogenerated.