Rename titan-yampa-recorder to titan-record-yampa. Refs #16.
For consistency with the module structure and to facilitate connecting other
implementations without stepping on each corresponding package's module tree,
this package is renamed to titan-record-yampa. Although it cannot be truly
interpreted that way because it provides reactimate replacements, the intuition
behind it is that record can be understood as the function and yampa as the
specific backend.