Add missing modules to cabal file. Refs #27.
The version of titan released on hackage does not include all modules in titan,
because the modules are not listed in the cabal file. Although these modules
need to be listed to create the dist tarball, they are not strictly necessary
to compile and install the package locally.

This change lists all modules in src/ (except for main) in the other-modules
section of the titan executable in the cabal file.