Introduce version of writeFile that flushes the handle. Refs #29.
The standard version of writeFile is too lazy. This is causing even simple apps
not to record correctly, because they do so at every loop iteration and the
file is not being closed before being open again for writing.

This change flushes the handle right after writing to it and before closing it.
This causes the full evaluation of the string written to it, and also that the
file is closed before it may be open again.