Remove unused modules. Refs #31.
Some of the modules included in the repository (which are now in the cabal
file, since #27) are not actually being used by titan.  This is preventing it
from installing straight from hackage, since some versions of cabal (but not
all) try to compile also those additional modules, which are out of date.

This change eliminates any module that is not used by titan, both the file
and the mention in the cabal file.