Rename hlint main test to avoid name clash on MacOS. Refs #30.
This repo contains two files with similar names in tests/, one uppercase and
one lowercase. The difference between uppercase and lowercase names is frequent
on linux, but not on mac or windows. This would make it very hard to use these
projects on those architectures, and would confuse cabal and git.

This commit renames the main HLint test file to HLintMain.hs, and adapts the
cabal file accordingly.